miden_processor::crypto

Module hash

Structs

Blake3_256

256-bit output blake3 hasher.

Keccak256

Keccak256 hash function

Poseidon2

Implementation of the Poseidon2 hash function with 256-bit output.

Rpo256

Implementation of the Rescue Prime Optimized hash function with 256-bit output.

Rpx256

Implementation of the Rescue Prime eXtension hash function with 256-bit output.

Sha256

SHA-256 hash function.

Sha512

SHA-512 hash function.

Type Aliases

Blake3Digest

Alias for the generic Digest type, for consistency with other hash modules.