miden_processor::event

Trait EventHandler

pub trait EventHandler:
    Send
    + Sync
    + 'static {
    // Required method
    fn on_event(
        &self,
        process: &ProcessorState<'_>,
    ) -> Result<Vec<AdviceMutation>, EventError>;
}

An EventHandler defines a function that that can be called from the processor which can read the VM state and modify the state of the advice provider.

A struct implementing this trait can access its own state, but any output it produces must be stored in the process’s advice provider.

Required Methods

fn on_event( &self, process: &ProcessorState<'_>, ) -> Result<Vec<AdviceMutation>, EventError>

Handles the event when triggered.