miden_processor::trace

Module chiplets

Structs

Ace

Arithmetic circuit evaluation (ACE) chiplet.

AuxTraceBuilder

Constructs the execution trace for chiplets-related auxiliary columns (used in multiset checks).

Chiplets

This module manages the VM’s hasher, bitwise, memory, arithmetic circuit evaluation (ACE) and kernel ROM chiplets and is responsible for building a final execution trace from their stacked execution traces and chiplet selectors.

ChipletsTrace

CircuitEvaluation

Contains the variable and evaluation nodes resulting from the evaluation of a circuit. The output value is checked to be equal to 0.

MerkleRootUpdate

Result of a Merkle tree node update. The result contains the old Merkle_root, which corresponding to the old_value, and the new merkle_root, for the updated value. As well as the row address of the execution trace at which the computation started.

Constants

MAX_NUM_ACE_WIRES

PTR_OFFSET_ELEM

PTR_OFFSET_WORD