miden_processor

Trait DebugHandler

pub trait DebugHandler: Sync {
    // Provided methods
    fn on_debug(
        &mut self,
        _process: &ProcessorState<'_>,
        _options: &DebugOptions,
    ) -> Result<(), DebugError> { ... }
    fn on_trace(
        &mut self,
        _process: &ProcessorState<'_>,
        _trace_id: u32,
    ) -> Result<(), TraceError> { ... }
}

Handler for debug and trace operations

Provided Methods

fn on_debug( &mut self, _process: &ProcessorState<'_>, _options: &DebugOptions, ) -> Result<(), DebugError>

This function is invoked when the Debug decorator is executed.

The default implementation is a no-op.

fn on_trace( &mut self, _process: &ProcessorState<'_>, _trace_id: u32, ) -> Result<(), TraceError>

This function is invoked when the Trace decorator is executed.

The default implementation is a no-op.