miden_processor

Trait MastForestStore

pub trait MastForestStore {
    // Required method
    fn get(&self, procedure_hash: &Word) -> Option<Arc<MastForest>>;
}

A set of MastForests available to the prover that programs may refer to (by means of an miden_core::mast::ExternalNode).

For example, a program’s kernel and core library would most likely not be compiled directly with the program, and instead be provided separately to the prover. This has the benefit of reducing program binary size. The store could also be much more complex, such as accessing a centralized registry of MastForests when it doesn’t find one locally.

Required Methods

fn get(&self, procedure_hash: &Word) -> Option<Arc<MastForest>>

Returns a MastForest which is guaranteed to contain a procedure with the provided procedure hash as one of its procedure, if any.